requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9
pytest>=8.2.0
httpx>=0.28.0
PyMuPDF>=1.24.0
python-docx>=1.1.2
python-dotenv>=1.0.1
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0
pymilvus>=2.6.0
neo4j>=6.0.0
